libtraceevent: Fix a double free in process_op()
When process_cond() failed, it freed the token but didn't reset the
arg->op.op to NULL.  So it tried to free the arg->op.op again from
free_arg() from the caller and resulted in a double free.

And Steve found another location that needs an update.  Let's handle
the error in consolidate_op_arg() instead.

Link: https://lore.kernel.org/linux-trace-devel/20240719163242.690f4c0e@gandalf.local.home/
Link: https://lore.kernel.org/linux-trace-devel/20240722161917.991077-1-namhyung@google.com

Cc: Ian Rogers <irogers@google.com>
Fixes: 76a0eb8d5a20c ("libtraceevent: Fix event-parse memory leak in process_cond")
Suggested-by: Steven Rostedt <rostedt@goodmis.org>
Signed-off-by: Namhyung Kim <namhyung@google.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>