libtraceevent: Add support for \t char in print fmt
Events can have \t characters in their print format strings. An example
is the smb3_cmd_done event from the cifs module. It has print format
string:
	 "\tsid=0x%llx tid=0x%x cmd=%u mid=%llu", ...

Add \t as an accepted character in the while loop that parses the format
string. This prevents the "[FAILED TO PARSE]" message and imporperly
parsed output in `trace-cmd report`.

Link: https://lore.kernel.org/linux-trace-devel/20250227202618.54073-1-devosruben6@gmail.com

Link: https://bugzilla.kernel.org/show_bug.cgi?id=219688
Signed-off-by: Ruben Devos <devosruben6@gmail.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>