libtraceevent: Have static library to hide hidden functions
When the static library uses the objects directly to be created, it makes
all the hidden functions visible to applications that link to it.

Check all the functions that are global and if they do not start with
"tep_" rename the symbol to start with "_tep_local_". This keeps the
source code clean and still adds the namespace for the hidden functions
within the objects.

Cc: Douglas Raillard <douglas.raillard@arm.com>
Tested-by: Metin Kaya <metin.kaya@arm.com>
Reviewed-by: Metin Kaya <metin.kaya@arm.com>
Link: https://lore.kernel.org/20251112144912.268258-2-rostedt@goodmis.org
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>