libtraceevent: Have meson build rename hidden global functions
The meson build needs to perform the same operation as the normal make
build with respect to global hidden functions within the static library.

Move the redefine_local function to utils.mk and have it take a parameter
for what shared library it will read to find the global functions with.
Then separate out a target in Makefile.meson for the static library to use
this helper function to rename the non "tep_*" global functions to have a
"_tep_local_*" prefix.

Cc: Douglas Raillard <douglas.raillard@arm.com>
Tested-by: Metin Kaya <metin.kaya@arm.com>
Reviewed-by: Metin Kaya <metin.kaya@arm.com>
Link: https://lore.kernel.org/20251112144912.268258-3-rostedt@goodmis.org
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>