libtraceevent: Have BTF find functions with extra characters
Sometimes the symbol of a function has extra characters followed by a '.'.
In this case, the BTF lookup may not find the function. If the function is
not found and the function name contains a '.', try stripping off the
extra characters and try again. That way, instead of:

  __mutex_unlock_slowpath.isra.0(ffffffffb5cd4e40, 8, 1, 1, ffffffffb43e0fe4, ffffed1024699f88)

It can produce:

  __mutex_unlock_slowpath.isra.0(lock=0xffffffffb5cd4e40, ip=0x8)

Cc: Douglas Raillard <douglas.raillard@arm.com>
Cc: Masami Hiramatsu <mhiramat@kernel.org>
Cc: Namhyung Kim <namhyung@kernel.org>
Cc: Takaya Saeki <takayas@google.com>
Cc: Ian Rogers <irogers@google.com>
Link: https://lore.kernel.org/20251112214929.290686-4-rostedt@goodmis.org
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>