libtraceevent: Do not change names of functions not of this library
The hidden attribute is used for functions within the library that are
not to be exported. They do not have name space naming which is not a
problem for the shared library, because the hidden attribute hides them
from being linked.

But these functions are not protected by the hidden attribute for the
static library. The build was modified to run 'nm' on the shared library
and any hidden function found that does not start with 'tep_' would have
its name changed to 'tep_local_<name>'.

A problem occurred where on x86_32 systems, gcc adds other functions like:

  __stack_chk_fail_local
  __udivdi3
  __umoddi

and these too were converted. But by doing so, it caused the "make test"
build to fail (make test uses the static library):

  /usr/bin/ld: lib/libtraceevent.a(event-parse.o): in function `consolidate_op_arg':
  src/event-parse.c:2302: undefined reference to `_tep_local___udivdi3'
  /usr/bin/ld: src/event-parse.c:2305: undefined reference to `_tep_local___umoddi3'
  /usr/bin/ld: lib/libtraceevent.a(event-parse.o): in function `eval_num_arg':
  src/event-parse.c:4561: undefined reference to `_tep_local___udivdi3'
  /usr/bin/ld: src/event-parse.c:4564: undefined reference to `_tep_local___umoddi3'
  /usr/bin/ld: lib/libtraceevent.a(event-parse.o): in function `print_event_time':
  src/event-parse.c:7168: undefined reference to `_tep_local___udivdi3'
  /usr/bin/ld: src/event-parse.c:7175: undefined reference to `_tep_local___umoddi3'
  /usr/bin/ld: src/event-parse.c:7175: undefined reference to `_tep_local___udivdi3'
  collect2: error: ld returned 1 exit status
  make[1]: *** [Makefile:22: utest/trace-utest] Error 1
  make: *** [Makefile:258: test] Error 2

Read the 'nm' symbols of the temporary static library and record only the
functions that have an address. As the above gcc functions are still
undefined there. Then this will only convert functions that are defined by
the static library and not added by the compiler.

Also add the missing .gitignore file to hide these temp files.

Cc: Douglas Raillard <douglas.raillard@arm.com>
Cc: Metin Kaya <metin.kaya@arm.com>
Link: https://lore.kernel.org/20251230103000.30be6e94@gandalf.local.home
Reported-by: Vitaly Chikunov <vt@altlinux.org>
Closes: https://lore.kernel.org/all/d3ozbooqkmvjf4rnyklwvdvc7h@altlinux.org/
Fixes: 399f8462ff1b1 ("libtraceevent: Have static library to hide hidden functions")
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>